能够跑通整个流程，但三角形还是没有绘制出来。需要找一下问题所在

+++ KYEngine/Graphic/DX/Dx11StateObject.cpp
@@ -49,7 +49,7 @@ namespace KY
 		d3dRTDesc.DestBlendAlpha= dx11Trans->ToBlendType(rtDesc.alphaOpDesc.dst);
 		d3dRTDesc.BlendOpAlpha	= dx11Trans->ToBlendOperation(rtDesc.alphaOpDesc.op);
 
-		d3dRTDesc.RenderTargetWriteMask = dx11Trans->ToColorWriteEnable(rtDesc.renderTargetWriteMask);
+		d3dRTDesc.RenderTargetWriteMask = uint8(dx11Trans->ToColorWriteEnable(rtDesc.renderTargetWriteMask));
 	}
 
 	bool DX::Dx11BlendStateObj::Create(const BlendState &state)
